AQLP-642 - consolidate tests
